[REFACTOR] gitignore 정리 + main.py 간결하게 정리
- 통합 실행 코드인 main.py가 깔끔한 게 보기 좋을 것 같아서 기존에 안에 있었던 transform 함수를 crawler/transform_comments.py 안으로 옮겨 main.py를 간소화 했습니다.
